+++ .github/actions/5_builderpackage_indexer_branch_resolver/resolve_branches.sh
@@ -6,10 +6,12 @@ set -e
 REPOS=(
     "wazuh-indexer-plugins"
     "wazuh-indexer-reporting"
+    "wazuh-indexer-security-analytics"
 )
 REPO_URLS=(
     "https://github.com/wazuh/wazuh-indexer-plugins.git"
     "https://github.com/wazuh/wazuh-indexer-reporting.git"
+    "https://github.com/wazuh/wazuh-indexer-security-analytics.git"
 )
 
 declare -A BRANCH_EXISTS