+++ build-scripts/baptizer.sh
@@ -12,6 +12,7 @@ function usage() {
     echo -e "-r REVISION\t[Optional] Package revision, default is '0'."
     echo -e "-l PLUGINS_HASH\t[Optional] Commit hash from the wazuh-indexer-plugins repository"
     echo -e "-e REPORTING_HASH\t[Optional] Commit hash from the wazuh-indexer-reporting repository"
+    echo -e "-s SECURITY_HASH\t[Optional] Commit hash from the wazuh-indexer-security-analytics repository"
     echo -e "-m MIN\t[Optional] Use naming convention for minimal packages, default is 'false'."
     echo -e "-x RELEASE\t[Optional] Use release naming convention, default is 'false'."
     echo -e "-h help"
@@ -22,7 +23,7 @@ function usage() {
 # ====
 function parse_args() {
 
-    while getopts ":hp:a:d:r:l:e:mx" arg; do
+    while getopts ":hp:a:d:r:l:e:s:mx" arg; do
         case $arg in
         h)
             usage
@@ -44,9 +45,10 @@ function parse_args() {
             PLUGINS_HASH=$OPTARG
             ;;
          e)
-
             REPORTING_HASH=$OPTARG
-
+            ;;
+        s)
+            SECURITY_HASH=$OPTARG
             ;;
         m)
             IS_MIN=true
@@ -139,8 +141,8 @@ function get_devel_name() {
         PREFIX="$PREFIX"-min
     fi
     # Generate composed commit hash
-    if [ -n "$PLUGINS_HASH" ] && [ -n "$REPORTING_HASH" ]; then
-        COMMIT_HASH="$GIT_COMMIT"-"$PLUGINS_HASH"-"$REPORTING_HASH"
+    if [ -n "$PLUGINS_HASH" ] && [ -n "$REPORTING_HASH" ] && [ -n "$SECURITY_HASH" ]; then
+        COMMIT_HASH="$GIT_COMMIT"-"$PLUGINS_HASH"-"$REPORTING_HASH"-"$SECURITY_HASH"
     fi
     PACKAGE_NAME="$PREFIX"_"$VERSION"-"$REVISION"_"$SUFFIX"_"$COMMIT_HASH"."$EXT"
 }